Refactoring from "paths" to "baseUrl"
