+++ src/pages/learn/index.tsx
@@ -1,4 +1,4 @@
-import {getTags} from '@lib/tags'
+import {getTags} from 'lib/tags'
 import Link from 'next/link'
 import {FunctionComponent} from 'react'
 import {GetServerSideProps} from 'next'
